Fixed some compiler dependent bugs in problems 54 and 76

+++ Euler.h
@@ -102,7 +102,7 @@ int32		problem73(int32 n);				// Number of reduced proper fractions greater than
 int32		problem74(int32 n, int32 m);	// Number of digit factorial chains with exactly m elements starting a number below n.
 int32		problem75(int32 n);				// Number of integer sided right triangles with a unique perimeter at most n.
 
-int32		problem76(int32 n);				// Number of different ways to write n as a sum of at least two positive integers.
+int64		problem76(int32 n);				// Number of different ways to write n as a sum of at least two positive integers.
 int32		problem77(int32 n);				// First number that can be written as the sum of primes in over n different ways.
 int32		problem78(int32 n);				// Smallest number that can be partitioned in a number of ways divisible by n.
 string		problem79();					// Find the shortest possible passcode from the ordered digits.
